+++ src/util.cpp
@@ -159,6 +159,8 @@ void Util::ParseArgs(const std::vector<std::string>& cargs,
   app.add_flag("--sora-disable-signaling-url-randomization",
                config.sora_disable_signaling_url_randomization);
   app.add_option("--sora-channel-id", config.sora_channel_id, "Channel ID");
+  app.add_option("--sora-client-id", config.sora_client_id, "Client ID");
+  app.add_option("--sora-bundle-id", config.sora_bundle_id, "Bundle ID");
   app.add_option("--sora-role", config.sora_role, "Role")
       ->check(CLI::IsMember({"sendonly", "recvonly", "sendrecv"}));
 
@@ -516,6 +518,8 @@ std::vector<std::vector<std::string>> Util::NodeToArgs(const YAML::Node& inst) {
       DEF_BOOLEAN(sora, "sora-", "disable-signaling-url-randomization");
 
       DEF_STRING(sora, "sora-", "channel-id");
+      DEF_STRING(sora, "sora-", "client-id");
+      DEF_STRING(sora, "sora-", "bundle-id");
       DEF_STRING(sora, "sora-", "role");
       DEF_BOOLEAN(sora, "sora-", "video");
       DEF_BOOLEAN(sora, "sora-", "audio");